share the guard that lets a test catch a FATAL_ERROR
FATAL_ERROR raises SIGTERM through exit_kotekan() before it throws, so a
test checking that throw has to handle the signal first. test_hdf5N2Write
carried its own guard for that; move it next to configure() in
test_logging.hpp, where the opposite handler already lives, so the next
test needing one does not copy it again.
